+++ core/src/net/oschina/j2cache/util/FstSnappySerializer.java
@@ -16,7 +16,6 @@
 package net.oschina.j2cache.util;
 
 import java.io.IOException;
-import java.io.Serializable;
 
 import org.xerial.snappy.Snappy;
 
@@ -42,12 +41,12 @@ public class FstSnappySerializer implements Serializer {
 	}
 	
 	@Override
-	public byte[] serialize(Serializable obj) throws IOException {
+	public byte[] serialize(Object obj) throws IOException {
 		return Snappy.compress(inner.serialize(obj));
 	}
 
 	@Override
-	public Serializable deserialize(byte[] bytes) throws IOException {
+	public Object deserialize(byte[] bytes) throws IOException {
 		if (bytes == null || bytes.length == 0)
 			return null;
 		return inner.deserialize(Snappy.uncompress(bytes));