+++ spring-boot-starter/src/net/oschina/j2cache/cache/support/redis/SpringRedisCache.java
@@ -8,16 +8,15 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import net.oschina.j2cache.Level2Cache;
 import org.springframework.data.redis.core.RedisTemplate;
-
-import net.oschina.j2cache.redis.RedisCache;
 
 /**
  * 
  * @author zhangsaizz
  *
  */
-public class SpringRedisCache implements RedisCache {
+public class SpringRedisCache implements Level2Cache {
 
 	private String namespace;
 
@@ -55,7 +54,7 @@ public class SpringRedisCache implements RedisCache {
 	}
 
 	@Override
-	public Serializable get(String key) {
+	public Object get(String key) {
 		Object value = redisTemplate.opsForHash().get(region, getKeyName(key));
 		if (value == null) {
 			return null;
@@ -64,8 +63,8 @@ public class SpringRedisCache implements RedisCache {
 	}
 
 	@Override
-	public Map<String, Serializable> getAll(Collection<String> keys) {
-		Map<String, Serializable> map = new HashMap<>(keys.size());
+	public Map<String, Object> get(Collection<String> keys) {
+		Map<String, Object> map = new HashMap<>(keys.size());
 		for (String k : keys) {
 			Object value = redisTemplate.opsForHash().get(region, getKeyName(k));
 			if (value != null) {
@@ -83,13 +82,13 @@ public class SpringRedisCache implements RedisCache {
 	}
 
 	@Override
-	public void put(String key, Serializable value) {
+	public void put(String key, Object value) {
 		redisTemplate.opsForHash().put(region, getKeyName(key), value);
 	}
 
 	@Override
-	public void putAll(Map<String, Serializable> elements) {
-		Map<String, Serializable> map = new HashMap<>(elements.size());
+	public void put(Map<String, Object> elements) {
+		Map<String, Object> map = new HashMap<>(elements.size());
 		elements.forEach((k, v) -> {
 			map.put(getKeyName(k), v);
 		});
@@ -126,4 +125,8 @@ public class SpringRedisCache implements RedisCache {
 		return (byte[]) redisTemplate.opsForHash().get(region, key);
 	}
 
+	@Override
+	public void setBytes(String key, byte[] bytes) {
+		redisTemplate.opsForHash().put(region, key, bytes);
+	}
 }