+++ CHANGES.md
@@ -1,5 +1,10 @@
 # J2Cache 版本更新记录
 
+**J2Cache 2.3.0-beta (2018-1-15)**
+* 内部结构的重构，减少模块间的耦合，简化内部接口
+* 去掉对 `commons-beanutils` 的依赖，由于 beanutils 1.8 和 1.9 版本 api 变化大，此举可避免对宿主系统的影响
+* 修复了 RedisGenericCache 中 clear 方法的逻辑错误[BUG]
+
 **J2Cache 2.2.4-release (2018-1-14)**
 * 修复了配置中设置 Caffeine 缓存有效期时，单位无效的问题